Refactor internal amqp-1.7.2 nr package
