+++ instrumentation/rabbit-amqp-1.7.2/src/test/java/com/nr/agent/instrumentation/rabbitamqp172/RabbitMQTest.java
@@ -5,7 +5,7 @@
  *
  */
 
-package com.nr.agent.instrumentation.rabbitamqp27;
+package com.nr.agent.instrumentation.rabbitamqp172;
 
 import com.newrelic.agent.introspec.InstrumentationTestConfig;
 import com.newrelic.agent.introspec.InstrumentationTestRunner;
